requirements.txt
asgiref==3.5.0
boto3==1.21.32
botocore==1.24.32
Django==4.0.3
django-filter==21.1
django-storages==1.12.3
gunicorn==20.1.0
jmespath==1.0.0
mysqlclient==2.1.0
Pillow==9.1.0
python-dateutil==2.8.2
python-dotenv==0.20.0
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.0.0
